feat: mkae List component adaptable for unordered list
