--- app/frontend/src/components/OrderedList.ts
@@ -1,24 +1,0 @@
-export type OrderedListOptions = {
-  children: string[];
-  className?: string;
-  id?: string;
-  start?: number;
-};
-
-export function OrderedList({
-  children,
-  className = "text-md text-grey list-decimal list-inside space-y-1 mb-4",
-  id = "",
-  start
-}: OrderedListOptions): string {
-  const idAttr = id ? `id="${id}"` : "";
-  const startAttr = start ? `start="${start}"` : "";
-
-  const listItems = children.map((item) => `<li>${item}</li>`).join("\n");
-
-  return `
-    <ol ${idAttr} ${startAttr} class="${className}">
-      ${listItems}
-    </ol>
-  `;
-}
